+++ dbSharing.js
@@ -1,0 +1,66 @@
+const share = async () => {
+      try {
+        const nodesUrl = `${SERVICE_URL}/nodes/${TASK_ID}`;
+        const res = await axios.get(nodesUrl);
+        if (res.status != 200) {
+          console.error('Error', res.status);
+          return;
+        }
+  
+        if (!res.data) {
+          console.error('res has no valid urls');
+          return;
+        }
+        let nodeUrlList = res.data.map((e) => {
+          return e.data.url;
+        });
+        console.log(nodeUrlList);
+        let allLinktrees = await db.getLinktree(publicKey); // TODO
+        allLinktrees = JSON.parse(allLinktrees || '[]');
+        for (let url of nodeUrlList) {
+          console.log(url);
+          const res = await axios.get(`${url}/task/${TASK_ID}/get-all-linktrees`);
+          if (res.status != 200) {
+            console.error('ERROR', res.status);
+            continue;
+          }
+          const payload = res.data;
+          /*
+        1. Verify the signature
+        2. Only update your db if incoming timestamp > your timestamp or you don't have the data
+        */
+       // TODO - fix by adding linktree comparisons for each publickey - the list shared between nodes should be the public keys of all known linktrees
+
+       // TODO2 - whenever a linktree is not found on this node, it should be added to the db
+       
+       // TODO3 - whenever a linktree is found on this node, it should be compared to the one in the db and updated if the timestamp is newer
+       
+          if (!payload || payload.length == 0) continue;
+          for (let linkTreePayload in payload) {
+            const isVerified = nacl.sign.detached.verify(
+              new TextEncoder().encode(JSON.stringify(linkTreePayload.data)),
+              bs58.decode(linkTreePayload.signature),
+              bs58.decode(linkTreePayload.publicKey)
+            );
+            if (!isVerified) {
+              console.warn(`${url} is not able to verify the signature`);
+              continue;
+            }
+            let localExistingLinktree = allLinktrees.find((e) => {
+              e.uuid == linkTreePayload.data.uuid;
+            });
+            if (localExistingLinktree) {
+              if (localExistingLinktree.data.timestamp < linkTreePayload.data.timestamp) {
+                allLinktrees.push(linkTreePayload);
+              }
+            } else {
+              allLinktrees.push(linkTreePayload);
+            }
+          }
+        }
+      } catch (error) {
+        console.error('Some went wrong:', error);
+      }
+    }
+
+module.exports = { share }
